requirements.txt
apache-beam[gcp]
requests
google-cloud-storage
